+++ CHANGES.rst
@@ -22,6 +22,7 @@ Changelog
 - Added new optional field (decisionSuite) for item
 - In tests WF helpers (validateItem, decideMeeting, ...) added parameter as_manager, True by default for MeetingItem related methods and False
   by default for Meeting related methods.  This way we avoid as much as possible hidden permission problems
+- Exclude SearchableText indexing for IAnnex objects
 
 4.1.19.2 (2020-03-17)
 ---------------------